Add feature to notify plugins of number of 'missed' lines of text
A line in 'missed' if it was filtered out due to a matchrule that does
not have that plugin set as a destination.

This can be useful when debugging and it seems that some messages are
not coming through. It is easy to forget that there is a rule that is
filtering out that message. So now you can see the missed number
increasing to be reminded of this situation.